+++ common/tools/dev-tool/src/commands/run/bundle.ts
@@ -70,9 +70,12 @@ export default leafCommand(commandInfo, async (options) => {
 
     try {
       const bundle = await rollup.rollup(baseConfig);
-
+      const cjsFilename = info.packageJson.main;
+      if (!cjsFilename) {
+        throw new Error("Expecting valid main entry");
+      }
       await bundle.write({
-        file: "dist/index.js",
+        file: cjsFilename,
         format: "cjs",
         sourcemap: true,
         exports: "named",