+++ sdk/policyinsights/arm-policyinsights/src/policyInsightsClient.ts
@@ -62,7 +62,7 @@ export class PolicyInsightsClient extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-policyinsights/6.0.0-beta.2`;
+    const packageDetails = `azsdk-js-arm-policyinsights/6.0.0-beta.3`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`